VR: fix compiler error

+++ Assets/Scripts/Game/Loader.cs
@@ -342,8 +342,8 @@ public class Loader : MonoBehaviour {
 						Globals.treesBelowGround = false;
 				}
 
-				if (xn ["rewardDur"] != null) {  // Used by SDT task where reward size is varied, but if the stock reward size is used, than a 2-4x multiplier is too much and the water drop falls onto the ball...
-					float.TryParse (xn ["rewardDur"].InnerText, out Globals.rewardDur);
+				if (xn ["rewardDur"] != null) {  // Used by SDT task where reward size is varied, but if the stock reward size is used (60ms or 4ul), than a 2-4x multiplier is too much and the water drop falls onto the ball...
+					int.TryParse (xn ["rewardDur"].InnerText, out Globals.rewardDur);
 				}
             }
